Add NSTextView-based editor to demo app
SwiftUI's TextEditor doesn't support rich text at all, not even varying text color. So we cannot use it as a source code editor with syntax highlighting.

Known bug in our new NSTextView wrapper: the syntax highlighting doesn't update when you edit the text in the NSTextView. It only updates when an outside change comes in, e.g. by editing the text in the SwiftUI TextEditor.